% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square_Plot_Functions.R
\name{delta_beta_plot}
\alias{delta_beta_plot}
\title{Create subfigures for Figure 3}
\usage{
delta_beta_plot(data, element)
}
\arguments{
\item{data}{}

\item{element}{a string denoting a TE subfamily name}
}
\value{
a ggplot object
}
\description{
Create subfigures for Figure 3
}
