% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square_Plot_Functions.R
\name{methy_plot}
\alias{methy_plot}
\title{Create subfigures for Figure 3}
\usage{
methy_plot(data, element)
}
\arguments{
\item{data}{}

\item{element}{a string denoting a TE subfamily name}
}
\description{
Create subfigures for Figure 3
}
