% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square_Plot_Functions.R
\name{plot_DMP_density}
\alias{plot_DMP_density}
\title{Create subfigures for Figure 3}
\usage{
plot_DMP_density(data, element)
}
\arguments{
\item{data}{}

\item{element}{a string denoting a TE subfamily name}
}
\value{
a ggplot object
}
\description{
Create subfigures for Figure 3
}
