% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square_Plot_Functions.R
\name{plot_normal_tumor_beta}
\alias{plot_normal_tumor_beta}
\title{Create subfigures for Figure 3}
\usage{
plot_normal_tumor_beta(data, element)
}
\arguments{
\item{data}{}

\item{element}{a string denoting a TE subfamily name}
}
\value{
a ggplot2 object
}
\description{
Create subfigures for Figure 3
}
