% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complex_Heatmap_Plot_Functions.R
\name{plotCorHeat}
\alias{plotCorHeat}
\title{Create subfigures for Figure 3}
\usage{
plotCorHeat(data, star)
}
\arguments{
\item{data}{}

\item{star}{}
}
\description{
Create subfigures for Figure 3
}
